+++ src/Transaction.h
@@ -8,12 +8,12 @@ using namespace std;
 class Transaction {
 	int idUser;
 	int idTransaction;
-	float value;
+	double value;
 	string name;
 	string issueDate;
 
 public:
-	Transaction(int idUser=0, int idTransaction=0, float value=0, string name="", string issueDate="2000-01-01") {
+	Transaction(int idUser=0, int idTransaction=0, double value=0, string name="", string issueDate="2000-01-01") {
 		this->idUser = idUser;
 		this->idTransaction = idTransaction;
 		this->value = value;
@@ -24,14 +24,14 @@ public:
 	//Container setters.
 	void setIdUser(int newIdUser);
 	void setIdTransaction(int newIdTransaction);
-	void setValue(float newValue);
+	void setValue(double newValue);
 	void setName(string newName);
 	void setIssueDate(string newIssueDate);
 
 	// Container getters.
 	int getIdUser();
 	int getIdTransaction();
-	float getValue();
+	double getValue();
 	string getName();
 	string getIssueDate();
 };
